requirements.txt
azure-cognitiveservices-speech==1.31.0
azure-storage-blob==12.18.3
pandas==2.1.4
numpy==1.26.2
soundfile==0.12.1
pylint==3.0.3
python-dotenv==1.0.0